requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.7.0
pytest-html==4.1.1
pytest-xdist==3.6.1
allure-pytest==2.13.5
python-dotenv==1.0.1
openpyxl==3.1.3
pandas==2.2.3
